Use lazy loading to split up the size of the js downloaded
